make cert script bug fix

+++ root/tls/certificate.sh
@@ -2,7 +2,7 @@
 #
 # using this script to generate certificate which signed by ca
 
-# target host name, such as ssl.example.com, must specified
+# target host names, such as ssl.example.com, split up by `,`, must specified
 TARGET_HOST=$1
 
 # output location, must specified
@@ -18,8 +18,8 @@ CA_PROFILE=$4
 
 # generate ca
 generateca(){
-	# generate ca-config
-	cat >$CA_LOCATION/ca-config.json <<-EOF  
+  # generate ca-config
+  cat >$CA_LOCATION/ca-config.json <<EOF
 {
   "signing": {
     "default": {
@@ -38,11 +38,11 @@ generateca(){
     }
   }
 }
-	EOF
+EOF
 
-	cat >$CA_LOCATION/ca-csr.json <<-EOF  
+  cat >$CA_LOCATION/ca-csr.json <<EOF
 {
-  "CN": "certificate",
+  "CN": "kubernetes",
   "key": {
     "algo": "rsa",
     "size": 2048
@@ -52,63 +52,79 @@ generateca(){
       "C": "China",
       "ST": "BeiJing",
       "L": "BeiJing",
-      "O": "certificate",
-      "OU": "ca"
+      "O": "k8s",
+      "OU": "System"
     }
   ]
 }
-	EOF
+EOF
 
-	cfssl gencert -initca $CA_LOCATION/ca-csr.json | cfssljson -bare $CA_LOCATION/ca
+  cfssl gencert -initca $CA_LOCATION/ca-csr.json | cfssljson -bare $CA_LOCATION/ca
 }
 
 # check ca whether existed
 checkca(){
-	if [[ -z "$CA_LOCATION" ]]; then
-		CA_LOCATION="~/.ca"
-	fi
+  if [[ -z "$CA_LOCATION" ]]; then
+    CA_LOCATION="~/.ca"
+  fi
 
-	if [[ ! -d "$CA_LOCATION" ]]; then
-		mkdir -p $CA_LOCATION
-	fi
+  if [[ ! -d "$CA_LOCATION" ]]; then
+    mkdir -p $CA_LOCATION
+  fi
 
-	# no ca file, generate ca
-	if [[ ! -f "$CA_LOCATION/ca.pem" ]]; then
-		generateca
-	fi
+  # no ca file, generate ca
+  if [[ ! -f "$CA_LOCATION/ca.pem" ]]; then
+    generateca
+  fi
 
-	if [[ -z "$CA_PROFILE" ]]; then
-		CA_PROFILE="certificate"
-	fi
+  if [[ -z "$CA_PROFILE" ]]; then
+    CA_PROFILE="certificate"
+  fi
 }
 
 #
 # BEGIN SHELL
 # check target
 if [[ -z "$TARGET_HOST" ]]; then
-	echo "`TARGET_HOST` not specified, exit"
-	exit -1
+  echo "`TARGET_HOST` not specified, exit"
+  exit -1
 fi
+
+IFS=","
+TARGET_HOST_ARRAY=($TARGET_HOST)
+TARGET_HOSTS=""
+
+for host in ${TARGET_HOST_ARRAY[@]}; do
+  TARGET_HOSTS=${TARGET_HOSTS}'"'${host}'",'
+done
+TARGET_HOSTS_LEN=${#TARGET_HOSTS}-1
+# TARGET_HOSTS=${TARGET_HOSTS:0:TARGET_HOSTS_LEN}
+echo $TARGET_HOSTS
 
 # check whether specified location of output
 if [[ -z "$OUTPUT_LOCATION" ]]; then
-	echo "`OUTPUT_LOCATION` not specified, exit"
-	exit -1
+  echo "`OUTPUT_LOCATION` not specified, exit"
+  exit -1
 fi
 if [[ ! -d "$OUTPUT_LOCATION" ]]; then
-	mkdir -p $OUTPUT_LOCATION
+  mkdir -p $OUTPUT_LOCATION
 fi
 
 # check ca
 checkca
 
-# generate target-csr.json
-cat >$OUTPUT_LOCATION/target-csr.json <<EOF
+# generate cert-csr.json
+cat >$OUTPUT_LOCATION/cert-csr.json <<EOF
 {
-    "CN": "$TARGET_HOST",
+    "CN": "kubernetes",
     "hosts": [
       "127.0.0.1",
-      "$TARGET_HOST"
+      $TARGET_HOSTS
+      "kubernetes",
+      "kubernetes.default",
+      "kubernetes.default.svc",
+      "kubernetes.default.svc.cluster",
+      "kubernetes.default.svc.cluster.local"
     ],
     "key": {
         "algo": "rsa",
@@ -116,11 +132,11 @@ cat >$OUTPUT_LOCATION/target-csr.json <<EOF
     },
     "names": [
         {
-            "C": "-",
-            "ST": "-",
-            "L": "-",
-            "O": "$TARGET_HOST",
-            "OU": "$TARGET_HOST"
+            "C": "China",
+            "ST": "Beijing",
+            "L": "Beijing",
+            "O": "system",
+            "OU": "system"
         }
     ]
 }
@@ -135,7 +151,11 @@ cfssl gencert -ca=$CA_LOCATION/ca.pem \
               | cfssljson -bare $OUTPUT_LOCATION/cert
 
 # formate certificate file into base64
-cat $OUTPUT_LOCATION/cert-key.pem | base64 | xargs echo -n | sed -e 's/[ ][ ]*//g' >$OUTPUT_LOCATION/cert-key.pem.base64
-cat $OUTPUT_LOCATION/cert.pem     | base64 | xargs echo -n | sed -e 's/[ ][ ]*//g' >$OUTPUT_LOCATION/cert.pem.base64
+if [[ -f "$OUTPUT_LOCATION/cert-key.pem" ]]; then
+  cat $OUTPUT_LOCATION/cert-key.pem | base64 | xargs echo -n | sed -e 's/[ ][ ]*//g' >$OUTPUT_LOCATION/cert-key.pem.base64
+fi
+if [[ -f "$OUTPUT_LOCATION/cert.pem" ]]; then
+  cat $OUTPUT_LOCATION/cert.pem | base64 | xargs echo -n | sed -e 's/[ ][ ]*//g' >$OUTPUT_LOCATION/cert.pem.base64
+fi
 
 exit 0
